+++ tests/transforms/test_transforms.py
@@ -30,11 +30,13 @@ def sample() -> Dict[str, Tensor]:
 def batch() -> Dict[str, Tensor]:
     return {
         "image": torch.tensor(  # type: ignore[attr-defined]
-            [[[[1, 2, 3], [4, 5, 6], [7, 8, 9]]]]
-        ).to(torch.float),
+            [[[[1, 2, 3], [4, 5, 6], [7, 8, 9]]]],
+            dtype=torch.float,  # type: ignore[attr-defined]
+        ),
         "mask": torch.tensor(  # type: ignore[attr-defined]
-            [[[[0, 0, 1], [0, 1, 1], [1, 1, 1]]]]
-        ).to(torch.long),
+            [[[[0, 0, 1], [0, 1, 1], [1, 1, 1]]]],
+            dtype=torch.float,  # type: ignore[attr-defined]
+        ),
         "labels": torch.tensor([[0, 1]]),  # type: ignore[attr-defined]
     }
 
@@ -87,11 +89,13 @@ def test_identity(sample: Dict[str, Tensor]) -> None:
 def test_augmentation_sequential(batch: Dict[str, Tensor]) -> None:
     expected = {
         "image": torch.tensor(  # type: ignore[attr-defined]
-            [[[[3, 2, 1], [6, 5, 4], [9, 8, 7]]]]
-        ).to(torch.float),
+            [[[[3, 2, 1], [6, 5, 4], [9, 8, 7]]]],
+            dtype=torch.float,  # type: ignore[attr-defined]
+        ),
         "mask": torch.tensor(  # type: ignore[attr-defined]
-            [[[[1, 0, 0], [1, 1, 0], [1, 1, 1]]]]
-        ).to(torch.long),
+            [[[[1, 0, 0], [1, 1, 0], [1, 1, 1]]]],
+            dtype=torch.long,  # type: ignore[attr-defined]
+        ),
         "labels": torch.tensor([[0, 1]]),  # type: ignore[attr-defined]
     }
     augs = transforms.AugmentationSequential(
